Added automation check/fix for Aruba SNMP issue.
This is a new script to automate the check for and fix of
a bug in Aruba switch FW that causes SNMP to be missing
MAC address information that is available (and correct)
in the switch management CLI.  This causes HMS discovery
to not see some hardware and consequently not discover it.

This script automates the check for whether the issue is currently
affecting the system, and if so, resets the SNMP server on the
Aruba switches to fix the problem.

NOTE: This is not 100% hands-off.  If Aruba switches exhibit
the problem and require the SNMP reset, the script will perform it,
but the user has to enter the Aruba switch admin password.  It can't
be automated for security reasons.

DOES THIS CHANGE INVOLVE ANY SCHEME CHANGES?  N

REMINDER: HAVE YOU INCREMENTED VERSION NUMBERS? E.G., .spec, Chart.yaml N/A

REMINDER 2: HAVE YOU UPDATED THE COPYRIGHT PER hpe GUIDELINES: © Copyright 2014-2020 Hewlett Packard Enterprise Development LP    ? Y

* Resolves CASMHMS-5059

Tested on:

* baldar

Was a fresh Install tested? N   No charts or docker images involved.
Was an Upgrade tested?      N   " "
Was a Downgrade tested?     N   " "
Was a CT test run?          N   " "
If schema changes were part of this change, how were those handled in your upgrade/downgrade testing? N/A

On baldar, the steps that the script performs were done manually to
verify that the discovery/SNMP issue was actively happening.

The script was run, and the actions were verified to have been executed.
The next HSM discovery did indeed see the BMCs that had been missing,
verifying that the script fixed the problem.

Very low risk.