CASMINST-5858: Rework verify_hsm_discovery.py to support the EX2500 cabinet (#35)
Reworked the verify_hsm_discovery.py script to properly supports the EX2500 cabinet. There now exists a single validation function that supports all cabinets types, instead of duplicating logic for each cabinet type. This is useful as the EX2500 cabinet can contain both river and mountain hardware.

Added some quality of life fixes to reduce the amount of noise in the validation output, to make the interpretation of the output simpler.

Only report missing mountain/hill Nodes,NodeBMCs, and RouterBMCs if the slot is actually populated. This is required as SLS assumes fully populated cabinets due to not having a good source of data for this. So only populated slots are considered, and empty slots are ignored. This is a major nuisance if a cabinet is not fully populated as the output of the script will print a line for each node that doesn't actually exist in the system.
SLS assumes all nodes look like windom nodes when assigning NIDs and creating components. For nodes like Bardpeak/Grizzly Peak/Antero blades where they have a different configuration of node BMCs and nodes on each blade, and false positive warnings are detected for the components that SLS has that don't actually exist on the blade. So these phantom components are ignored in the validation output if the model of blade can be determined, otherwise the old behavior is used.
For example on grizzly peak blades only have 1 BMC b0 and 2 nodes b0n0 b0n1, and the components b1, b1n0 and b1n1 do not exist.
The root issue for this is going to be addresses in CASM-3111
Don't emit a warning for the BMC of ncn-m001 missing from HSM, as it is currently expected to not be connected to the HMN network, and therefore cannot be discovered.
Don't emit a warning for "Intel CMCs" as those peices of hardware do not really exist in real life.