CASMTRIAGE-7692 adjust `run_hms_ct_tests.sh` to reduce false positives (#63)
when running via goss

this changes the `else` condition of this wrapper script to use an
alternate log file name to prevent false failures when running this
script in parallel.

the main use case here is that the goss tests from csm-testing runs this
script in several processes when called via goss, and they all try to
read and write to the same log file.  this works ok when run serially,
such as in a simple `for` loop, but when they are all kicked off at
once, it can lead to odd situations like this completed log file:

```
NAME: cray-power-control
LAST DEPLOYED: Wed Aug 28 07:49:37 2024
NAMESPACE: services
STATUS: deployed
REVISION: 7
TEST SUITE:     cray-power-control-test-functional
Last Started:   Wed Jan 15 15:44:56 2025
Last Completed: Wed Jan 15 15:46:25 2025
Phase:          Succeeded
TEST SUITE:     cray-power-control-test-smoke
Last Started:   Wed Jan 15 15:44:24 2025
Last Completed: Wed Jan 15 15:44:56 2025
Phase:          Succeeded
         Succeeded
ray-hms-smd:

SUCCESS: Service test passed: sls

SUCCESS: Service test passed: bss

SUCCESS: Service test passed: capmc

SUCCESS: Service test passed: hsm

SUCCESS: Service test passed: fas

SUCCESS: Service test passed: pcs
```

text is truncated unintentionally and not all output is present.

this script is a bit fraigle in this sense, since we determine success
or failure based on the output that is in this log file.  while this is
a helpful wrapper for the customer because it shows a nice summary of
output, but is less helpful for simple machines needing to check the
exit status or just print the `helm` output.

there is not a lot of impact to the manual execution path, except that
is creates a new log file if testing a single service.  so preventing
false positives in the ct pipeline is worth it.

during testing, I tried different fixes including a file lock and other
methods to not read/write to the same log file.  I ended up with this
solution since the code was close together, making a review of it
easier, and it also minimizes the footprint of the change.

the script hasn't had much work done for years, so I opted for this path
instead of trying to re-work the entire thing and disrupting other
places this script may be used.

Signed-off-by: Jacob Salmela <jacob.salmela@hpe.com>